% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridDensity.R
\name{gridDensity}
\alias{gridDensity}
\title{Create Grid Heat Map}
\usage{
gridDensity(data, shape, cs = c(1, 1), col = "blue", scale = TRUE,
  heat = FALSE, main = "Grid Density Heat Map")
}
\arguments{
\item{data}{Spatial Points Data Frame object of Data}

\item{shape}{Spatial Polygon Data Frame}

\item{cs}{Size of grid}

\item{main}{Title for Map}

\item{factor}{Categorical Variable to Plot}

\item{time}{Title for Map}

\item{legend}{Boolean for whether to plot legend}

\item{size}{Size of subplots}

\item{legend.pos}{Position of legend}
}
\value{
geospatial visualization
}
\description{
\code{gridDensity} Create Grid Heat Map
}
\details{
Create Grid Heat Map
}
\examples{
\dontrun{
gridDensity(df, shape)
}

}
