% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatMap.R
\name{heatMap}
\alias{heatMap}
\title{Plots cloroploth map of points}
\usage{
heatMap(data, shape = NULL, col = "blue", main = "Sample HeatMap")
}
\arguments{
\item{data}{Spatial Points Data Frame object of Data}

\item{shape}{Spatial Polygon Data Frame}

\item{col}{Color of heatMap}

\item{main}{Title for Map}
}
\value{
geospatial visualization
}
\description{
\code{heatMap} Plots cloropleth map
}
\details{
Plots chloropleth map
}
\examples{
\dontrun{
heatMap(df, shape)
}

}
