% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPieGrid.R
\name{plotPieGrid}
\alias{plotPieGrid}
\title{Plots Pie Charts over Grids}
\usage{
plotPieGrid(data, shape1, factor, main = "Pie Plot Over Grid", size = c(1,
  1), cs = c(5, 5), legend = TRUE, legend.pos = "bottomright")
}
\arguments{
\item{data}{Spatial Points Data Frame object of Data}

\item{factor}{Categorical Variable to Plot}

\item{main}{Title for Map}

\item{size}{Size of subplots}

\item{cs}{Size of grid}

\item{legend}{Boolean for whether to plot legend}

\item{legend.pos}{Position of legend}

\item{shape}{Spatial Polygon Data Frame}
}
\value{
geospatial visualization
}
\description{
\code{plotPieGrid} Plots Pie Charts over Grids
}
\details{
Superimposes Pie Charts over Grids
}
\examples{
\dontrun{
plotPieGrid(df, shape, factor)
}

}
