% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPiePoly.R
\name{plotPiePoly}
\alias{plotPiePoly}
\title{Plots Pie Charts over Polygons}
\usage{
plotPiePoly(data, shape, factor, main = "Pie Plot", size = c(1, 1),
  legend = TRUE, legend.pos = "bottomright")
}
\arguments{
\item{data}{Spatial Points Data Frame object of Data}

\item{shape}{Spatial Polygon Data Frame}

\item{factor}{Categorical Variable to Plot}

\item{main}{Title for Map}

\item{size}{Size of subplots}

\item{legend}{Boolean for whether to plot legend}

\item{legend.pos}{Position of legend}
}
\value{
geospatial visualization
}
\description{
\code{plotPiePoly } Plots Pie Charts over Polygons
}
\details{
Superimposes Pie Charts over polygons
}
\examples{
\dontrun{
plotPiePoly (df, shape, factor)
}

}
