% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendlinePolygon.R
\name{trendlinePolygon}
\alias{trendlinePolygon}
\title{Plots Trendlines over Polygons}
\usage{
trendlinePolygon(data, shape, Breaks = "months", time, main = "Plot Title",
  normal = TRUE, size = c(1, 1), Color = palette())
}
\arguments{
\item{data}{Spatial Points Data Frame object of Data}

\item{shape}{Spatial Polygon Data Frame}

\item{time}{Title for Map}

\item{main}{Title for Map}

\item{normal}{Boolean for whether to normalize points}

\item{size}{Size of subplots}

\item{Color}{palette}

\item{breaks}{Color of heatMap}
}
\value{
geospatial visualization
}
\description{
\code{trendlinePolygon} Plots Trendlines over Polygons
}
\details{
Superimposes trendline plots over polygons
}
\examples{
\dontrun{
heatMap(df, shape)
}

}
